requirements.txt
torch
torchvision
diffusers>=0.19.0
transformers
accelerate
safetensors
bitsandbytes
datasets
peft
Pillow
tqdm
